feat: refacor ProCard to remove circular dependency
