+++ packages/card/src/index.tsx
@@ -1,8 +1,9 @@
 import ProCard from './Card';
+import type { ProCardProps } from './Card';
 import StatisticCard from './components/StatisticCard';
 import Statistic from './components/Statistic';
 
-import type { ProCardTabsProps, ProCardProps } from './Card';
+import type { ProCardTabsProps } from './type';
 import type { StatisticCardProps, StatisticsCardProps } from './components/StatisticCard';
 import type { StatisticProps } from './components/Statistic';
